requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
tavily-python==0.5.0
firecrawl-py==1.5.0
python-dotenv==1.0.0
pydantic-settings==2.6.0
motor==3.6.0
pymongo==4.9.1
pydantic[email]==2.10.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
